Simplify Showtime component, accept single child
Previously the Showtime component took a function as a child,
which took a ref and attached it to the wrapped component. That
function is now moved to be inside the library, simplifying the
client API.

The way the ref is attached to the child element is by cloning
the element and attaching the ref to the clone. This may have
some performance penalty, depending on the complexity of the
component. This may also have other consequences, say if the
client code is tracking that component otherwise (e.g. by using
a ref of their own).